backend/admin-service/requirements.txt
asgiref==3.8.1
Django==4.2.16
djangorestframework==3.15.2
mysqlclient==2.2.6
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
django-cors-headers==3.14.0
prometheus-client==0.19.0
requests==2.31.0
django-prometheus==2.3.1
python-dotenv==1.0.0
